Separate function resolution and completion
This changeset divides the full resolution of functions into two steps.
In the first step (handled by resolve_decl_function) the function
prototype is resolved and a symbol for the function is added to the
global symbol table. Then some metadata describing the resolved function
is added to a list of incomplete functions (to be completed after the
resolution of all top-level declarations). In the second step (handled
by complete_function) the body of an incomplete function is resolved,
fully preparing that function for the code generation phase.

With this separation of function prototype resolution and function body
resolution, mutually recursive functions may be resolved without running
into a dependency error as both functions will have access to the full
list of function-symbols in the global symbol table.