Add byte literals
Currently conversion to the byte type is performed by an implicit cast
from u8 or s8 expressions. Moving forward I would like to remove
implicit casts from the language, and doing so requires giving bytes
proper literals.

This changeset gives byte literals the y suffix. A suffix of b clashes
with the hex digit so y is used as it is the next character in "byte"
after b.