Add const slice expressions
Allow for the evaluation of slice expressions in a constant evaluation
context.

This changeset also fixes some of the ugly code used to generate static
data in NASM, replacing the existing system with a more elegant and less
buggy string builder pattern. Additional portions of this changeset
address issues with code generation related to indexing and lvalue
addressing for zero-sized objects.