Forbid use of global variables in const evaluation
This changeset fixes a bug in the evaluator where any symbol with a
compile-time-known initial value could be evaluated in a constant
expression. This bug allowed constants to be initialized with the
initial value of a global variable. An explicit check has been added in
the evaluator, causing a fatal error if a non-constant non-function
symbol is evaluated a constant expression.