Fix unaligned copy in array rvalue index operation
Array rvalues placed on the stack are pushed with a total padded size
that aligns them to a modulo 8 address. The copy operation on the
indexed element was using the array size instead of this padded size
when calculating the copy destination address. This changeset adjusts
the offset calculation to use the correct (padded) value.