Check for invalid rvalue array LHS in slice expression
The resolve phase was not producing an error when an rvalue array was
used as the left-hand-side of a slicing operation. This would cause an
assertion to trigger during code generation as there is no stable
address for the start of an rvalue array. This changeset adds a check
for rvalue arrays in the slice expression resolve function.