Fix incorrect is-lvalue return for <expr-index>
The tir_expr_is_lvalue function was incorrectly reporting that an index
expression with a left-hand-side of slice type could not be used as an
lvalue even though code generation for such a scenario already existed
in codegen.c. This changeset updates the tir_expr_is_lvalue function to
correctly report that an index expression with left-hand-side slice may
be used in an lvalue or rvalue index expression.