Add countof operator
The unary countof operator takes a right-hand-side expression of array
or slice type and returns the number of indexable elements in that array
or slice as a usize. For right-hand-side expressions of array type the
countof expression produces the statically-known count of the array. For
right-hand-side expressions of slice type the countof expression
produces the count word internal to the slice data structure. Side
effects of the right-hand-side expression are observed.