Use explicit register sizes for cmp instructions
Commit 065ef395dae7e7ef2a8999eb6e6617dfdb73ab19 involved ~5 hours of
debugging due to an error in the way that a cmp instruction on two 8-bit
values was using 64-bit registers, accidentally comparing the garbage
data in the upper seven bytes of each register. This commit updates
other operations using the cmp instruction to avoid this same problem.

The code generation phase is kind of a mess right now so there might be
some comparisons that were missed.