Fix redeclaration error from two global symbols in separate namespaces
I encountered the following error when adding a `read` function under
the `std` namespace in lib/std/io.sunder even though the previous
declaration of `read` in lib/sys/sys.sunder was under the `sys`
namespace.

----
[/home/ashn/sources/sunder/lib/std/io.sunder:8] error: redeclaration of `read` previously declared at [/home/ashn/sources/sunder/lib/sys/sys.sunder:198]
func read(fd: ssize, buf: []byte) ssize {
^
----

It turns out that the error was caused by a bug in the resolve phase
that was adding top-level declarations to the global symbol table and
module export table for every module even though modules that declare a
namespace should only be exporting *that* namespace as a symbol.

A check was added that skips adding top-level declarations to the global
symbol table and module export table if the module being resolved
declares a namespace.