Fix incorrect code generation of namespaced function rvalues in constant expressions
Attempting to use a namespaced function as an rvalue in a global
variable declaration or constant declaration resulted in incorrect
assembly being generated. The function's local name was being used as
the assembly label rather than the function's static address name.

This changeset updates the code generation phase to use the function's
static address name for dx data generation instead of the function's
local name. The address of the function has been added to the
tir_function data structure to facilitate this change.