Improve performance when compiling one dimensional array literal constant expressions that use ellipsis initialization
Adds an ellipsis member to the value data structure that is used in
conjunction with NASM's times prefix in generated assembly to avoid
redundantly producing repeated ellipsis elements.

This optimization allows us to avoid cost of allocating N values for an
N element array literal during the resolve phase. Additionally, using
the ellipsis element with NASM's times prefix allows us to write the
data of the ellipsis element once and let the assembler handle repeating
that data in static memory.

Unfortunately, the times prefix cannot be nested, so arrays of arrays
are not able to benefit from this optimization to the same extent that
one dimensional arrays can, as only the inner-most array(s) may use the
prefix. It is speculated that ellipsis initialization will mostly be
used for zero-initializing one dimensional buffers, so hopefully the
lack of optimisation on arrays of arrays will not be noticeable in
practice.

In my original benchmark I was seeing ~22 seconds of real time to build
the brainfuck example program:
----
$ ./misc/build-debug.sh clean install >/dev/null 2>&1
$ time sunder-compile examples/bf.sunder

real	0m22.120s
user	0m5.954s
sys	0m16.166s
----

With this changeset that time has been reduced to ~0.3 seconds:
----
$ ./misc/build-debug.sh clean install >/dev/null 2>&1
$ time sunder-compile examples/bf.sunder

real	0m0.289s
user	0m0.118s
sys	0m0.171s
----