Remove conditional enabling of the debug logging function
The debug logging function is intended to be used for quick and dirty
printf-debugging when hacking on the compiler, so having to change a
`#define` in order to enable the function is a bit overkill. This
changeset removes conditional compilation for the debug logging
function, altering the function to behave more like the error and fatal
logging functions.