Cleanup temporary files on nasm or ld error
Currently the compiler will not clean up temporary .asm and .o files if
either of the spawn calls to nasm or ld fail during code generation.
This changeset uses a goto-error strategy for ensuring that resources
created during the code generation phase are cleaned up in either the
success or error cases (when the -k flag is not provided).

Note that the return values of the `remove` calls are unchecked. As a
future improvement it may be worth adding error checking and reporting
to these calls for all cases except ENOENT.