Define convention for local labels in generated assembly
Local labels take the form:
        .__<TIR-node-type>_<unique-id>_<description>

The <TIR-node-type> is STMT for stmt nodes and EXPR for expr nodes. The
<unique-id> is generated for each node in the call to codegen_stmt,
codegen_rvalue, and codegen_lvalue. The <description> is used to denote
what section of TIR node generation is taking place as well as provide
the necessary labels for jumps.

For each stmt and expr node, the local labels:
        .__<TIR-node-type>_<unique_id>_bgn:
and
        .__<TIR-node-type>_<unique_id>_end:
are generated to denote the beginning and end of code generation for that
node. These labels are always present and may be used for jump targets.