Change occurences of typeof identifier to typeof_ in C sources
The GNUC typeof extension causes an identifier conflict with the typeof
identifier in the compiler's CST typeof node. When compiling with
-pedantic-errors it seems that this extension is disabled, but
attempting to build bin/sunder-compiler using a GNUC compiler with the
default C99 release flags:

    make bin/sunder-compile CC=clang

This changeset changes all typeof identifiers to typeof_ to avoid
conflicting with the GNUC extension.