Add member functions
Member functions take the form of function declarations defined within
the body of a struct declaration.

    struct foo {
        var x: u16;
        var y: u16;

        func init(x: u16, y: u16) foo {
            return (:foo){
                .x = x,
                .y = y
            };
        }

        func sum(self: *foo) u16 {
            return (*self).x + (*self).y;
        }
    }

Member functions may be accessed as static functions useing the ::
operator with the struct type as the left hand side argument and the
member function name as the right hand side argument.

    var val: foo = foo::init(0x1111, 0x2222);
    var sum: u16 = foo::sum(&val);

Currently, member functions may not be called on instances of a struct
type (e.g. val.sum() instead of baz::sum(&val) in the example above),
but this functionality is expected to be added in the near future.