Overhaul symbol grammar
A bug in the Sunder grammar was uncovered while implementing template
structs which required rewriting a bunch of data structures and
algorithms internal to the compiler's parse, order, and resolve phases.
The new system put in place defines symbols recursively (even though
they are implemented as a list within the compiler) instead of the hacky
four separate but incorrect definitions for a symbol that existed
before.