Interfaces MVP
This changeset introduces a model for the definition of interface
structs. An interface uses a ("pointer to vtable", "pointer to
instance") pair an type erases the "self" parameter of member functions
implementing the interface. This MVP required creating a nasty function
casting hack and is still quite verbose, so additional work is required
in order to improve the ergonomics and safety of interface creation and
use.

This changeset defines the standard library interfaces `std::writer` and
`std::formatter`.