Ergonomic improvements to std::formatter and std::writer
Adds wrapper functions to the std::formatter and std::writer interfaces
that allow a user to avoid having to manually access the interfaces'
vtables and pass the `object` parameter.

Also adds the std::writer::write_all member function that continuously
calls std::writer::write until either an error occurs or all of the data
is written.