Remove -k flag from examples.build.sh
This flag is not necessary for new users who are looking at example
programs. The extra .o and .asm files are more of a distraction than a
help. If a user desires to view the generated assembly or object file
then they may always compile the example themselves with the -k flag.