Standard library simplification overhaul
Attempt to reduce the number of "one-off" source files in the standard
library. The goal of this changeset is to consolidate the most common
data structures and algorithms into one module that every other
standard library module can import. Structs such as `std::result` and
`std::writer` are used so frequently that it makes the most sense to
keep them together.

This new module `lib/std/core.sunder` serves as this common module. For
something to be added to the core module it must:

1. Have no other dependencies besides those in `lib/sys/sys.sunder`.
2. Never dynamically allocate memory.