Place statement and expression location comment after *_bgn label
When debugging generated assembly I would always get confused by which
location comment corresponded to which labels. This changeset moves the
position of that comment so that generated constructs follow the
pattern:

        THING_bgn:
        <THING-location>
        <THING-code>
        THING_end:

With this change, someone looking at the generated assembly knows that
the entirety of an expression or statement is contained within the *_bgn
and *_end labels.