Remove build-time dependency on vendored autil library
This changeset merges the code from the vendored autil library into the
sunder utility code. Having the functions and data structures provided
by autil.h has been incredibly helpful throughout the development of the
compiler thus far. However, we are now at a point where it is worth
forking and customizing autil to fit the specific needs of the Sunder
project rather than the more general-purpose needs that autil.h was
originally designed to handle.