Add std::iterator interface and initial implementations
The std::iterator interface is based off of the C# cursor iterator
model. This model was chosen for Sunder iterators in order to allow
iteration over a sequence using a for-expression loop without requiring
boilerplate within the loop body. Additionally, this model was chosen in
order to allow laziness when advancing an iterator as is seen in
streaming iterator models.