Fix incorrect exit status returned by sunder-run
The exit status of `rm` was being returned by sunder-run instead of the
exit status of the invoked program. This changeset fixes the issue by
capturing the status of the invoked program before running `rm`, and
then explicitly exiting with the captured status at the end of the
sunder-run script.