Allow compiling without linking
When the -c flag is provided to sunder-compile, the compiler will
compile a sunder source file, along with all its dependencies, into an
object file without linking that object file into an executable. When
compiling with -c, the `_start` definition will not be added to the
generated assembly and the `main` function may be omitted from the
Sunder program.