Move assembly fatal handlers into generated code
This changeset replaces the builtin __fatal_* set of assembly functions
with a single __fatal function defined in lib/sys/sys.asm. The builtin
__fatal function takes a message and message length as its first and
second function arguments, respectively, and is used as the single
application exit point in the case of a fatal error. The existing
__fatal_* functions are now automatically generated within codegen.c and
forward to the sys.asm-defined __fatal.