Transform alias declarations from symbol aliasing to type aliasing
Originally, alias declarations were designed to allow mapping a name
onto an existing symbol. However, in practice alias declarations were
always used as a pseudo-typedef. The original form of alias declarations
mapped a name to a symbol and not a name to a type, so alias
declarations such as

    alias str = []byte;

would be invalid. This changeset alters the behavior of alias
declarations to make alias a true typedef, and fixes a bunch of parsing
bugs (as well as one unrelated constant expression evaluation bug) that
appeared from the resulting grammar change.