Remove the struct resolver current_symbol_name_prefix member variable
This member variable *should* have been used to process/handle full
symbol names (mostly for type names in error messages), but this
functionality was never actually implemented. In the future it would be
nice to add proper support for full symbol names, but for now the code
is being removed since it does not actually do anything.