Remove allocator error message constants
With the upcoming stable-ish release of Sunder, I don't know if it is a
good idea to commit to the error-matches-this-byte-string method of
error checking. Error constants have been removed for now, but *may* be
added again in the future once we have a better error handling plan in
place.