Revert "Make failing to check function call return values emit a warning"
This reverts commit ecfcf53d81ad7e0cdd2400a772ed52b9ec422943.

This is a "nice-to-have", but cases such as `std::hash_map::insert`
would require the line `var _ = map.insert(key, value);` for the
majority of inserts. This funcitonality would be better handled by
something similar to `go vet` which could check for unused return values
for functions returning an `std::result`.