Add support for higher half kernel mappings
It was discovered during the creation of a bare bones kernel that
registerless NASM instructions in 64-bit mode are not RIP–relative by
default. The default absolute instructions produced "relocation
truncated to fit: R_X86_64_32S" linker errors when attempting to mov or
push an address located outside of the 32-bit address space.

This changeset updates the default NASM instruction form to be
RIP-relative, and changes instances of:

    push addr

into

    lea rax, [addr]
    push rax