Clean up constant folding
Currently the Sunder AST is not exposed directly as a tree on loaded
modules, so it is difficult to write optimization passes using a tree
traversal. For now, constant folding has been explicitly marked in
places where it is used with `OPTIMIZATION`. The intention is to
eventually move these optimizations out into a separate phase.

Also fixes some recently introduced bugs stemming from the work on
explicit and implicit casting.