Remove the dump statement from the language and add it as a sys function
Removes the keyword and associated functionality. Replaces uses of
`dump` in tests with the new `sys::dump[[T]]` function. Many preveously
dependency-free tests now rely on importing the system library, but this
was the best step in order to keep the test behavior exercised without
completely rewriting over a hundred tests.