Switch from an array list to a hash set for interned string lookup
Profiling with perf showed that the compiler was spending a large amount
of time in the `intern` function, likely due to the linear traversal
over the list of interned strings, as well as the `strlen` and `memcmp`
calls used to check each traversed element for string equality. This
changeset replaces the array list with a hash set in order to provide
average O(1) lookup when performing interning.

Testing the hash set implementation with `time make examples check`
showed an approximate 25%-50% performance improvement over the existing
array list implementation.