Align static objects with the `align` macro in generated NASM code
Prior to this changeset, objects with static storage duration would be
aligned semi-randomly based on the instructions and other static objects
that happened to appear before the target object in generated NASM code.
This changeset enforces the proper alignment of these static objects by
placing an `align` macro just before the label of each static object.