Fix incorrect final string production within string_split
This changeset fixes an error that would cause the final string that
*should* be produced by string_split to be incorrectly ignored when the
separator string is larger than the input string. Thankfully,
string_split is only ever called within canonical_import_path to split a
string on occurrences of the separator ":", which does not appear to
trigger this bug in practice.