Eliminate heap allocating for every token
This changeset removes the `xalloc` call that was made for each token
produced by the lexer. Each call to `lexer_next_token` originally
returned a frozen heap allocated token, a design wich was based on the
thinking that all references to tokens should be valid for the entire
lifetime of the compiler without having to think too hard. It turned out
that the eventual form of `struct token` ended up being a POD type that
did not strongly benefit from having a stable address, so these
individual frozen heap allocations were not necessary.

As a knock-on effect of the token changes, uses of `struct
source_location const*` were changed to `struct source_location` as the
source location pointers were no longer referencing the stable memory
within heap allocated tokens.