Use the getdents system call for sys::getdents in C sys.h file
The sys.h implementation of sys::getdents was previously using the
linux_dirent structure with the getdents64 system call because the
ARM-based laptop used for C backend development did not support
SYS_getdents and I wanted to get something building, even if it was
wrong. The incorrect struct-syscall pairing predicable lead to garbage
data populating the buffer dirent buffer. This changeset fixes the error
for x64, but will likely not build on ARM.