Fix undefined wrapping negation behavior in the C backend
With this changeset, all 412 of the tests in the test suite pass using
the C backend! This backend still requires a *lot* of cleanup, and there
are surely edge cases that have not yet been encountered, but the
existing C code generation is in a good enough spot where it is worth
investigating ports to other architectures.