Conditionally link lib/sys/sys.sunder to an arch-specific sys.sunder file
Sunder does not currently support conditional compilation, so there is
no in-source way to change the constants associated with the POSIX
platform layer found in sys.sunder based on the current architecture.
This changeset introduces a workaround that allows the use of multiple
platform-specific sys.sunder files by symlinking the correct file at
build-time depending on the current architecture. This is not a scalable
long-term solution, but it should be enough to get sunder running on
AMD64 Linux machines.