Add basic platform-specific conditional imports
During a directory import, files of the form `<stem>.sunder` will be
skipped in favor of the platform-specific files:

+ <stem>.<arch>-<host>.sunder
+ <stem>.<arch>.sunder
+ <stem>.<host>.sunder

if any such files exist within the imported directory.

This changeset is an MVP, and specifically does *not* cover
non-directory conditional imports. The rules for conditional imports
currently only look at <arch> and <host> strings, but this features has
been designed specifically to allow files of the form
`<stem>.<tag-1>-<tag-2>-<etc>.sunder` containing arbitrary tags to be
implemented in the future if desired.