Remove unnecessary extra object zeroing in generated C code
Structs already have their padding zeroed since structs are declared in
the generated C with extra padding member variables that are always
zero-initialized. The extra explicit zeroing was not performing any
useful work, but *was* causing significant performance problems for
running code, presumably from work being performed twice as well as
optimization prevention due to the use of volatile in the memzero
function.