Use pipes when compiling with the C backend
This changeset attempts to minimize the amount of disk IO by piping from
the Sunder compiler frontend to the C compiler backend as well as piping
between phases of C compilation via the GCC/Clang `-pipe` option.

Testing on a WSL desktop machine and Pinebook Pro showed an approximate
5% performance increase when compiling the example programs and test
suite with `make examples check` on a build with `CC=clang` and
`CFLAGS='$(GNU_REL)'`. Performance gains for the average application
will likely be a little bit higher since a majority of the test programs
do not import the standard library.