Remove the NASM/YASM backend
This changeset removes the NASM/YASM backend from the compiler, leaving
the C backend as the singular way to compile Sunder programs and
libraries.

The Sunder project originally had a much smaller scope, targeting only
x86-64 Linux with no intention of communicating with external code
(apart from the Linux kernel). NASM was chosen as the compiler backend
since it would ensure that reasonable assembly *could* be generated for
all Sunder language constructs, and because `nasm` + `ld` allowed the
compiler to produce static binaries with little effort. NASM was
definitely the right choice at the start of the project, but as the
scope of Sunder has grown, the inconveniences of directly generating
assembly have added up to the point where it no longer makes sense to
support NASM and YASM as code generation backends.

In the future, additional backends may be re-introduced to the compiler.
GNU flavored C is more portable than assembly thanks to widespread
adoption of GCC and Clang, but it would be nice to have other options in
the event that there is a niche platform not supported by a GNU C
compiler. Additionally, compiling Sunder to C and then C to machine
language (or Wasm) means that there is a lot of redundant work being
performed to run the same program through two compilers. None of these
are high-priority issues at the moment, but it shows that there is room
for growth and improvement in the future.