Fix incorrect source location displayed for member function calls
The source location used to indicate the location of a member function
call error would inaccurately show the member access dot to the left of
the member function identifier in expressions with nested member access.
This changeset updates the chosen source location to be the identifier
of the member function being invoked. Additionally, this changeset makes
member function error reporting consistent with the principle of least
surprise for a handful of other member function error messages emitted
by the compiler