Avoid interning and buffing generated C during code generation
Prior to this commit, generated rvalue and lvalue expressions were
interned as an ad-hoc means of cstring garbage collection, relying on
the intern pool finalizer to clean up allocations. This changeset
updates rvalue and lvalue expression generation to heap-allocate each
generated expression string and append that string to a temporary string
list. The temporary string list is periodically (currently after every
statement is generated) iterated-over in order to de-allocate strings
that are no longer in use.

Additionally, this changeset replaces the `struct string*` output buffer
used during code generation with a `FILE*` so that generated C can be
streamed to the output file instead of being buffered entirely in
memory.

After testing, these alterations did not substantially affect compile
times or maximum resident set size. However, these changes are
architecturally cleaner that the previous interning and buffering
strategies, and may serve as a pathway to future streaming-only code
generation.