Add peephole optimization for code generation of byte arrays
Byte arrays created from `embed` statements can cause to multi-MB byte
sequences to be written as generated C when asset files are embedded in
an executable. Performance analysis of the Sunder compiler showed an
overwhelming percentage of frontend compilation time being spent within
`strgen_value` when compiling a project with ~7500 lines of Sunder code
and ~5MB of embedded asset files.

This changeset adds a peephole optimization within `strgen_value` which
uses the fixed stride of text emitted for each element of a byte array
to pre-allocate and efficiently fill the string buffer corresponding to
the static C array representing the byte array object.