Update the `clean` target to more accurately remove build artifacts
Previously, the `clean` target would remove files of the form `a.out*`
as a way to delete isntances of `a.out` and `a.out.tmp.c` generated by
`sunder-compile` with no specified output filename. Update the `clean`
target so that all files of the form `*.out` and `*.tmp.*` are removed,
covering the previous `a.out` and `a.out.tmp.c` cases as well as other
instances where `<OUT>.tmp.c` was generated for output file `<OUT>`.