Include the name mangling prefix in generated code for local variables
Local variables use a fabricated identifier in generated C which has the
potential to cause a name collision with a static object or function.
Even if this scenario is unlikely, we specifically choose to prefix C
identifiers corresponding to local variables with the name mangle prefix
as a preventative measure.