Disable -fsanitize=function checks within generated C function definitions
This changeset disables -fsanitize=function for function definitions in
generated C. Sunder allows implicit casting of a function with parameter
types and/or a return type of type `*T` to a function type where those
same parameter types and/or return type are of type `*any`. This form of
casting matches the Sunder definition of compatible types, in which all
pointers are considered to be word-sized integers with the same size and
alignment.