Allow type aliases within struct and union declarations
Permit use of the syntax:

        type T = sometype;

within struct and union declarations. Added to codify a specific pattern
where a constant declared with an anonymous type was used as a way to
pseudo-declare named helper types. What used to be accomplished via the
constant member declaration:

        let FOO: struct { var x: ssize; } = uninit;
        ...
        func bar() typeof(FOO) { ... }

can now be more directly written as:

        type foo = struct { var x: ssize; };
        ...
        func bar() foo { ... }