Allow taking the address of rvalue expressions
This changeset permits the application of the addressof operator to
rvalue expressions. The result of the evaluated rvalue expression is
stored in an opaque local variable within the symbol table corresponding
to the lexical scope where that expression was evaluated. The address of
that opaque symbol is produced as the result of the addressof
expression.

The changeset additionally leverages this mechanism as a way to allow
calling member functions on rvalue expressions, giving Sunder the
ability to achieve member function calling behavior similar to C++.