Update std::hash_map and std::hash_set member functions taking non-owning key parameters to use pass-by-reference
With the recent language addition that allows taking references to an
rvalue expression, using pass-by-reference for the key/element
parameters in std::hash_map and std::hash_set member functions is now
viable without needing to assign that parameter to a variable first.
This change finally allows the standard library to be consistent with
the behavior that functions taking non-owning values due so using
pass-by-reference while functions taking owning values do so using
pass-by-value.

This is a breaking change to the std::hash_map and std::hash_set APIs.