Inline system arithmetic operations when possible
Compiling Sunder applications with C optimization level -O0 was causing
performance problems in tight loops, due in part to the arithmetic
operation functions in sys.h having non-zero overhead when not inlined.

This changeset attempts to inline these arithmetic operation functions
to the best of the compiler's ability in order to improve runtime
performance when no other optimizations are applied. This inlining is
only applied when compiling with a GNU C compiler, as the arithmetic
operation functions for non-GNU C compilers are longer, contain more
checks/branches, and could potentially lead to further degraded
performance when inlined due to instruction cache bloat.