Add the ability to specify the underlying type of enumeration types
This changeset adds the ability for enum declarations and anonymous enum
types to have their underlying type explicitly specified.

        enum foo: u16 { A; B; } # Enum foo with underlying type u16.
        enum: s64 { C; D; } # Anonymous enum with underlying type s64.

These changes have been introduced to harmonize Sunder enums with the
enhancements to enumerations introduced in N3030 and added to the C
standard in C23. Note that unlike C23 enumerations, Sunder explicitly
does *not* support enumerations with `bool` or `byte` as the underlying
type (the Sunder equivalents to C23's `bool` and `char` types,
respectively). Support for enums with these underlying types *may* be
added in the future.