Add `underlying_type` as a symbol under enumeration types
This changeset adds the symbol `underlying_type` to enumeration types.
This symbol is a type alias for the underlying type of the enumeration.
For example, the following code will declare both `x` and `y` as
variables of type `u32`, the underlying type of `enum foo`:

        enum foo: u32 { A = 123; B = 456; }
        var x = (:foo::underlying_type)foo::A;
        var y: foo::underlying_type = uninit;

This language feature has been added to further support the work to
harmonize Sunder enums with the C23 enumeration enhancements added in
N3030. Specifically, this gives Sunder something similar to C++11's
`std::underlying_type_t<T>`.