Fix incorrect sign extension when casting byte values to integer types with larger bit-width
The representation of the Sunder `byte` type is defined to match the
target platform C `char` type, which may or may not be signed. If the C
`char` type is signed, then Sunder `byte` values may undergo sign
extension if those values would be represented as a negative signed
8-bit integer. For example the Sunder code:

        var a = 0x80y;
        var b = 0x2800 | (:u16)a;

would generate C code that would sign extend `a` from 0x80 to 0xFF80
when implicitly promoting the value of `a` before bitwise ORing that
value with 0x2800. The bitwise OR of 0x2800 and 0xFF80 is 0xFF80, so the
result stored in `b` would appear to have garbage data instead of the
expected result of 0x2880.

This changeset normalizes byte values to unsigned 8-bit integers when
performing cast operations so that promotion to a larger-width integer
type does not sign extend the most significant bits of the result with
ones.