Check for typeof-target circular dependencies when a symbol starts with typeof
Previously, symbols starting with `typeof` would not be ordered based on
the typeof symbol, leading to a scenario in which this code:

    let A = typeof(B)::X;
    let B: S = uninit;
    struct S { let X = sizeof(typeof(A)); }

would produce the error message:

    [test.sunder:1] error: use of undeclared identifier `B`
    let A = typeof(B)::X;
                   ^

even though B is defined at the top level of the module.

With this changeset, the `typeof` expression, is now used for the cycle
check.

    [sunder:1] error: circular dependency created by declaration of `A`
    let A = typeof(B)::X;
    ^
    info: declaration of `A` (line 1) depends on `B` (line 2)
    info: declaration of `B` (line 2) depends on `S` (line 3)
    info: declaration of `S` (line 3) depends on `A` (line 1)