Remove Dockerfile, update README, and make finishing GRAMMAR touch-ups
This changeset contains the last bit of cleanup for a Sunder release
that will (hopefully) be used for a series of retrospective and
educational blog posts on the development of the Sunder language,
compiler, and standard library.

The changeset removes the x64 Linux development Dockerfile and
associated Docker instructions in the project README. This Dockerfile
had gone unused for some time, and is more or less unnecessary now that
Sunder can build on Linux and macOS natively was well as on Windows via
WSL.

This changeset also makes some fixes to the language GRAMMAR file after
taking another grammar proof reading pass.